requirements-test.txt
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0
pytest-mock>=3.11.0
responses>=0.23.0
freezegun>=1.2.0
requests-mock>=1.11.0